Add a key to the each loop to fix bad retypes.
The orders screen was using the index as the id for the item row
and issuing retypes and possibly other actions every time an item
was added because all prior entries in the order would be updated.

Instead, key the rows by item id so that svelte correctly adds a
new row when new entries are scanned instead of updating all the
prior entries, avoiding the redundant retype problem.